Define segments at run time for association tests
Instead of reading a predefined segments file, the segments file
will be created at run time prior to starting association tests.
Currently the default segment length is used, but that will be made
into an argument to assoc.py. Remove generic segments file and code
that saved it to the pipeline directory.