Consolidate documentation
In the README, config parameters common to multiple association
test types were repeated for each type. This makes maintainng the
docs difficult and obscures the differences between the types of
tests. Common parameters are now shown only once in an appropriate
subsection of the docs. Also, realized the section on parallelization
was out of date.