Conditional analysis
Allow conditional analysis: using dosage of variants as covariates.
The getPhenotypes function in TopmedPipeline is reponsible for
reading genotypes from the gds file. The conditional_chrom parameter
specifies which GDS file to look in for the variants, currently all
conditional variants in an analysis must be on the same chromosome.