use separate_rows
The output of aggregateList* was making the allele.index column a
character if any variants had >2 alleles. That propagated to
the variantInfo data.frame from assocTestSeq having all columns
coerced to factors, which caused an error in getAssoc when we tried
to take the minimum value of position. Using separate_rows from
tidyr makes the code of aggregateList much more elegant, and
always returns allele.index as an integer.