add option for supplying user-defined variant weights
Variant weights are given in an RData file, with the column name
provided as a config parameter. This file might be the same one
as used to define aggregate units. A function in TopmedPipeline
creates an AnnotatedDataFrame by joining the user-provided table
to the variant info from the GDS, using either variant.id or
position and allele information.