Utilize MKL multithreading
If MKL is build with multithreading enabled, we can set
MKL_NUM_THREADS to control how many threads are used by MKL for
matrix calculations. This is useful in code that doesn't already
use multithreading, that does large matrix calculations, and that
is not already running in parallel (because using multiple cores
per segment will probably not improve the overall performance).
The null model performance may with multithreading, so enable
using mutiple cores/threads in that code.