write KING as GDS instead of RData
KING output is very large, so save it as GDS instead of RData.
This is not natively supported by SNPRelate, so add function to
write output object to a GDS file and read it back again as a
snpgdsIBDClass object. This allows selecting a subset of samples
before reading instead of after.