Show the actual values used as the outcome in the nullmod report
Instead of plotting the distribution of the original outcome
variable in the null model report, plot the distribution of the
actual outcome. If running a model with rank-normalize residuals,
the intermediate report shows the histogram of the original values
and the final report shows the histogram of the rank-normalized
residual values. To make this change, do not remove the "outcome"
variable from the intermediate null model before saving it.