Do not include inverse_normal config value in null model report
The inverse_normal config value was confusing here when producing
the intermediate null model report: even though it was set to
true, the intermediate model was not run with inverse normalized
values. However, displaying the config value might make people
think that it was inverse normalized. It's also more of an analysis
parameter instead of a null model parameter. Therefore, don't
display it in the null model report.