Color the PY histogram by case/control status for binary outcomes
When preparing the data, convert the outcome to a factor if the
analysis is a binary analysis. This makes the plots better for
binary outcomes. Add coloring by case/control status to the PY
histogram; the previous step was necessary to get this to work.
This has the benefit of making some of the other plots more
appropriate when the binary outcome is coded as 0/1.