Only show inv normal parameters when the outcome is transformed
The config file printed the rescale_variance parameter in the
first model fitting step, before the inverse normal transform
had happened. Change the config printing so that this parameter is
only shown when the report is generated for an inverse-transformed
model. Also explicitly add the "inverse_normal" parameter, since
it is not really true for the first, pre-transform model fit
(although it is true for the analysis overall).