Fix outcome histogram when rank normalizing
After commit 338499aa1d50ce256fa221a19436325ff4706168 that
reorganized plotting data prep, the histogram of the outcome
accidentally ended up showing the original outcome instead of the
transformed outcome for the inverse-normal transformed model
report. Update the data prep and plot so that the correct outcome
from the model is shown.