Use out_prefix instead of out_file in null_model config
Constructing the filenames is much easier when we use a prefix in
the config file instead of the output file name. It is also more
consistent with how the other scripts work.  Note that the example
config files for the null_model.py script already use out_prefix,
so they don't need to be changed.