Convert null model workingY element to a vector
Sometimes the workingY element is a vector, but other times it's a
matrix, for example when a binary outcome/binomial family model is
fit when there are no non-zero variance components. In these cases,
passing a matrix to the data.frame function does not create a
column with the specified name (in this case, "workingY"); instead,
it renames the column to whatever the column name of the matrix is.
Fix this (at least for now) by converting the single-column matrix
to a vector.