Fix null model bug with binary variance components going to zero
The null model report job failed if a binary/binomial family model's
variance components all went to zero. This is because the
varCompCI function in GENESIS crashes when given the null model
object. Temporarily fix this so the report runs by checking the
zeroFLAG element and skipping the variance component calculation if
all variance components are zero.