Remove null model string functions from TopmedPipeline
The GENESIS null model how has the model string, so the functions
in TopmedPipeline are no longer needed. Remove the functions from
the exports for the package. Since this was handled by roxygen,
and because the roxygen version changed, the other man pages
generated with roxygen were updated with the new version number.