Make density plot by group figure larger and remove separate legend
This plot can often have long legends. Previously the report had
a second "figure" solely to print the legend, and it was at the end
of the report. After moving this figure to midway through the
section in stead of at the end, the "legend" figure left a lot of
plain whitespace. Make the figure bigger so that it can accomodate
a larger legend when necessary.