Fix outcome vs covariates plots for inverse normal models
The outcome plotted in the outcome vs. covariates plots was the
original outcome, not the inverse-normal transformed outcome.
Clean up the part of the code that sets the outcome variable
names and set the outcome_numeric variable to the outcome from
the null model, not the outcome from the phenotype file.