Use cluster-agnostic environment variables in runRscript
runRscript.sh was written for use with SGE, using SGE-specific
environment variables. Modify the script to process either SGE or
Slurm environment variables by setting cluster-agnostic variables
at the beginning of the script based on the SGE variables or the
Slurm variables, depending on are set. Then use these cluster-agnostic
variables in the remainder of the script.

Note: it would probably be better to have the Cluster subclasses
in TopmedPipeline.py set the proper variables for the script, if
we refactored the classes following #55.