Support hold_array for Slurm
This uses the --dependency=aftercorr:x,x,x.. flag. This requires that
a job task finish successfully for the dependent task to start. For
consistency, change the dependency for regular jobs (not array holds)
to afterok: instead of afterany:. This casues jobs dependent on any
failed jobs to sit in the queue instead of failing.